requirements-dev.txt
astor
astunparse
dask[distributed]
mock-ssh-server
pyshd
pytest
pytest-runner
setuptools >=28
setuptools_scm
sphinx
